Removed network simulator, using column_validator instead.
